[objects] Make ScopeInfo hashable
This CL implements a `Hash` function for ScopeInfo based on position
information. If no position information is available, we fall back to
the type and number of context variables.

Note that this is far from ideal, especially when no position info is
available. But, the hash is only used to store scope-related debug
information in the `LocalsBlocklistCache` hash table. This table is
only ever filled on debug pauses or debug-evaluates, so we don't
care that much if we produce many hash collisions.

R=jarin@chromium.org, leszeks@chromium.org

Doc: https://bit.ly/chrome-devtools-debug-evaluate-design
Bug: chromium:1363561
Change-Id: I70b7f2702693e2d930ed0080506ed94ac44e9124
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3925434
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83484}