[csa] Don't pass in unused var into NameDictionaryLookup
There's a use of NameDictionaryLookup where the outparam
`var_name_index` is unused but is still passed in. This turned into a
gotcha when modifying NameDictionaryLookup due CSA variable merging.
Remove the gotcha by not passing it in when it's not going to be used.

Change-Id: I11b41f347d97f6a77fd6f4bfda66df87c46737da
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3935212
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83562}