[liveedit] Pass outer ScopeInfo when live editing sloppy eval'd scripts
The CL https://crrev.com/c/3932724 changed LiveEdit to replace the
ScopeInfo object of "unchanged" SFIs instead of updating the position
info in place. As a side-effect, we no longer set the outer ScopeInfo
on sloppy EVAL_SCOPEs.

This CL fixes that by properly passing the outer ScopeInfo along to
the compiler. It's also necessary to pass the outer ScopeInfo to the
reparse of the original script. Otherwise we'd get false positives in
the logic that detects outer scope changes for unchanged functions.

R=leszeks@chromium.org

Bug: chromium:1363561
Change-Id: Ia0bd8363cd835d7c7ffdc1d157c6657e6e2275c3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3937802
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83584}