[debug] Fix block lists for scopes with identical positions
This CL fixes a bug in the block list handling for debug-evaluate.
Specifically, we can't rely on the source position alone to find
scopes. We also need to take the scope type into account. This needs
to happen in two places:

  - When we try to find the closure scope in the re-parse
    result based on the function at the top of the stack.

  - When we try to find matching `ScopeInfo` objects from
    re-parsed scopes.

Note that the code alrady contains a fix for a similar issue in
the past where we found class scopes instead of member initializer
function scopes. Both share the exact same source position.

This CL generalizes that fix to check that the scope type matches
when looking for the closure scope.

R=jarin@chromium.org

Bug: chromium:1363561
Change-Id: I18d167f59c994ece85971273954d0e3c65c12b7a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3959915
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83757}