[maglev] Move TranslationArray building back into code-gen
TranslationArray building was split off from code-gen, as a separate
graph pass, so that it could be done on the background thread. Now that
code-gen assembly is on the background thread, we can merge the
TranslationArray building back into codegen. To keep the public
interface uncluttered, this patch adds a MaglevTranslationArrayBuilder
helper class which wraps this functionality up.

A side-effect of this is that we now need the LocalIsolate in the code
generator (for value reification). Take this opportunity to pass in the
LocalIsolate instead of the Isolate, and pass the Isolate just into
GenerateCode.

Bug: v8:7700
Change-Id: I9377f84840895147d6d524ca42b3adf5ce6055f7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3966189
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83792}