[maglev] Move BytecodeAnalysis to graph builder
Avoid performing BytecodeAnalysis on the main thread, by moving the
BytecodeAnalysis class and constructor to the graph builder.

To support this, remove CheckIsLoopPhiIfNeeded from the interpreter
frame state -- this was a DCHECK that is more awkward to wire up now
that the bytecode analysis is on the graph builder and not the
compilation unit, and the frame state stuff has been stable long enough
now that it's probably not worth re-wiring.

Bug: v8:7700
Change-Id: I210cb88f575ca34efb85a3bf88b9f0e2303d92ca
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3963037
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83796}